Abstract terms and conditions modal
This commit was missed on the previous PR - had it locally but never hit push.
