+++ drizzle/relations.ts
@@ -1,0 +1,3 @@
+import { relations } from "drizzle-orm/relations";
+import { } from "./schema";
+
