+++ guards/TermsGuard.tsx
@@ -1,81 +1,84 @@
 import React, { useEffect, useState } from 'react';
 import { useAuth } from '@/contexts/AuthContext';
 import { useSystem } from '@/db/powersync/system';
-import {
-  Modal,
-  View,
-  Text,
-  TouchableOpacity,
-  StyleSheet,
-  ScrollView,
-  Linking
-} from 'react-native';
-import { colors, sharedStyles, spacing } from '@/styles/theme';
-import { useTranslation } from '@/hooks/useTranslation';
-import { Ionicons } from '@expo/vector-icons';
+import { View } from 'react-native';
 import { userService } from '@/database_services/userService';
-import { LanguageSelect } from '@/components/LanguageSelect';
-import { language } from '@/db/drizzleSchema';
-
-type Language = typeof language.$inferSelect;
+import { TermsModal } from '@/components/TermsModal';
 
 export function TermsGuard({ children }: { children: React.ReactNode }) {
   const { currentUser, setCurrentUser, isLoading } = useAuth();
   const { powersync, supabaseConnector } = useSystem();
-  const [selectedLanguage, setSelectedLanguage] = useState<Language | null>(
-    null
-  );
-  const { t } = useTranslation(selectedLanguage?.english_name);
   const [termsModalVisible, setTermsModalVisible] = useState(false);
-  const [termsAccepted, setTermsAccepted] = useState(false);
-  const [isProcessing, setIsProcessing] = useState(false);
+  const [hasAcceptedTerms, setHasAcceptedTerms] = useState(false);
 
   // Check if the user has accepted the terms and conditions
   useEffect(() => {
-    if (!isLoading && currentUser && !currentUser.terms_accepted) {
-      setTermsModalVisible(true);
+    if (!isLoading && currentUser) {
+      // Set the initial state based on the user's terms acceptance
+      const hasAccepted = !!currentUser.terms_accepted;
+      setHasAcceptedTerms(hasAccepted);
+
+      // Only show the modal if terms haven't been accepted
+      if (!hasAccepted) {
+        setTermsModalVisible(true);
+      }
     }
   }, [currentUser, isLoading]);
 
   // Pause syncing if terms not accepted
   useEffect(() => {
-    if (!isLoading && currentUser && !currentUser.terms_accepted) {
+    if (!isLoading && currentUser && !hasAcceptedTerms) {
       // Pause syncing
       powersync.disconnect();
-    } else if (!isLoading && currentUser && currentUser.terms_accepted) {
+    } else if (!isLoading && currentUser && hasAcceptedTerms) {
       // Resume syncing if not connected
       if (!powersync.connected) {
         powersync.connect(supabaseConnector);
       }
     }
-  }, [currentUser, isLoading, powersync]);
+  }, [currentUser, isLoading, powersync, hasAcceptedTerms, supabaseConnector]);
 
   const handleAcceptTerms = async () => {
-    if (!currentUser || isProcessing) return;
+    if (!currentUser) return;
 
     try {
-      setIsProcessing(true);
+      console.log('Accepting terms...');
 
-      // Update user profile with terms acceptance
-      const updatedUser = await userService.updateUser({
-        id: currentUser.id,
-        terms_accepted: true,
-        terms_version: '1.0'
-      });
+      // Update the user's metadata in auth
+      const { error: authError } =
+        await supabaseConnector.client.auth.updateUser({
+          data: {
+            terms_accepted: true,
+            terms_version: '1.0'
+          }
+        });
 
-      if (updatedUser) {
-        setCurrentUser(updatedUser);
-        setTermsModalVisible(false);
+      if (authError) {
+        console.error('Error updating user metadata:', authError);
+        return;
+      }
 
-        // Resume syncing
-        if (!powersync.connected) {
-          powersync.connect(useSystem().supabaseConnector);
+      // Update local state to reflect terms acceptance
+      setHasAcceptedTerms(true);
+
+      // Close the modal
+      setTermsModalVisible(false);
+
+      // Resume syncing
+      if (!powersync.connected) {
+        powersync.connect(supabaseConnector);
+      }
+
+      // Refresh the current user
+      const { data: userData } = await supabaseConnector.client.auth.getUser();
+      if (userData.user) {
+        const updatedUser = await userService.getUserById(userData.user.id);
+        if (updatedUser) {
+          setCurrentUser(updatedUser);
         }
       }
     } catch (error) {
       console.error('Error accepting terms:', error);
-    } finally {
-      setIsProcessing(false);
     }
   };
 
@@ -85,207 +88,25 @@ export function TermsGuard({ children }: { children: React.ReactNode }) {
   }
 
   // If user has accepted terms or is not authenticated, render children
-  if (!currentUser || currentUser.terms_accepted) {
+  if (!currentUser || hasAcceptedTerms) {
     return <>{children}</>;
   }
 
   // Otherwise, show terms modal and block rendering of children
   return (
-    <Modal
-      animationType="slide"
-      transparent={true}
-      visible={termsModalVisible}
-      onRequestClose={() => {
-        // Only allow closing if not a new user (terms already accepted)
-        if (currentUser?.terms_accepted) {
-          setTermsModalVisible(false);
-        }
-      }}
-    >
-      <View style={styles.modalContainer}>
-        <View style={styles.modalContent}>
-          <View style={styles.modalHeader}>
-            <Text style={styles.modalTitle}>
-              {t('termsAndConditionsTitle')}
-            </Text>
-            <Text style={styles.modalVersion}>{t('termsVersion')}</Text>
-            <TouchableOpacity
-              style={styles.closeButton}
-              onPress={() => {
-                // Only allow closing if not a new user (terms already accepted)
-                if (currentUser?.terms_accepted) {
-                  setTermsModalVisible(false);
-                }
-              }}
-              disabled={!currentUser?.terms_accepted}
-            >
-              <Ionicons
-                name="close"
-                size={24}
-                color={
-                  currentUser?.terms_accepted
-                    ? colors.text
-                    : colors.textSecondary
-                }
-              />
-            </TouchableOpacity>
-          </View>
-
-          {/* Language Selector */}
-          <View style={styles.languageSelector}>
-            <LanguageSelect
-              value={selectedLanguage?.id || ''}
-              onChange={(lang) => setSelectedLanguage(lang)}
-              containerStyle={{ flex: 1 }}
-            />
-          </View>
-
-          <ScrollView
-            style={styles.modalBody}
-            contentContainerStyle={styles.modalBodyContent}
-          >
-            <Text style={styles.modalText}>
-              {t('termsAndConditionsContent')}
-            </Text>
-            <Text style={styles.modalText}>{t('termsContributionInfo')}</Text>
-            <Text style={styles.modalText}>{t('termsDataInfo')}</Text>
-            <TouchableOpacity
-              onPress={() => {
-                // Open the full data policy in browser
-                Linking.openURL(t('dataPolicyUrl'));
-              }}
-              style={{ marginTop: spacing.medium }}
-            >
-              <Text style={[sharedStyles.link, styles.linkText]}>
-                {t('viewFullDataPolicy')}
-              </Text>
-            </TouchableOpacity>
-          </ScrollView>
-          <View style={styles.termsCheckbox}>
-            <TouchableOpacity
-              onPress={() => setTermsAccepted(!termsAccepted)}
-              style={styles.checkboxContainer}
-            >
-              <Ionicons
-                name={termsAccepted ? 'checkbox' : 'square-outline'}
-                size={24}
-                color={colors.text}
-                style={styles.checkboxIcon}
-              />
-              <Text style={styles.checkboxText}>{t('agreeToTerms')}</Text>
-            </TouchableOpacity>
-          </View>
-          <View style={styles.modalFooter}>
-            <TouchableOpacity
-              style={[
-                sharedStyles.button,
-                { flex: 1 },
-                !termsAccepted && styles.disabledButton
-              ]}
-              onPress={handleAcceptTerms}
-              disabled={!termsAccepted || isProcessing}
-            >
-              <Text style={sharedStyles.buttonText}>
-                {isProcessing ? t('processing') : t('accept')}
-              </Text>
-            </TouchableOpacity>
-          </View>
-        </View>
-      </View>
-    </Modal>
+    <>
+      <TermsModal
+        visible={termsModalVisible}
+        onClose={() => {
+          // Only allow closing if terms have been accepted
+          if (hasAcceptedTerms) {
+            setTermsModalVisible(false);
+          }
+        }}
+        onAccept={handleAcceptTerms}
+        showAcceptButton={true}
+        canDismiss={hasAcceptedTerms}
+      />
+    </>
   );
 }
-
-const styles = StyleSheet.create({
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    padding: 20
-  },
-  modalContent: {
-    backgroundColor: colors.background || 'white',
-    padding: 20,
-    borderRadius: 20,
-    width: '100%',
-    maxWidth: 500,
-    maxHeight: '90%',
-    alignItems: 'center'
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    width: '100%',
-    marginBottom: 15
-  },
-  modalTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: colors.text,
-    flex: 1
-  },
-  modalVersion: {
-    fontSize: 14,
-    color: colors.textSecondary
-  },
-  languageSelector: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    width: '100%',
-    marginBottom: 15,
-    gap: 10
-  },
-  languageLabel: {
-    fontSize: 16,
-    color: colors.text
-  },
-  modalBody: {
-    width: '100%',
-    maxHeight: '60%'
-  },
-  modalBodyContent: {
-    padding: 10
-  },
-  modalText: {
-    color: colors.text,
-    marginBottom: 15,
-    fontSize: 16,
-    lineHeight: 24
-  },
-  linkText: {
-    fontSize: 16,
-    textAlign: 'center'
-  },
-  termsCheckbox: {
-    width: '100%',
-    marginTop: 20,
-    marginBottom: 10
-  },
-  checkboxContainer: {
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    gap: spacing.small
-  },
-  checkboxIcon: {
-    marginTop: 2
-  },
-  checkboxText: {
-    color: colors.text,
-    flex: 1,
-    flexWrap: 'wrap'
-  },
-  modalFooter: {
-    flexDirection: 'row',
-    width: '100%',
-    marginTop: 10
-  },
-  disabledButton: {
-    opacity: 0.5
-  },
-  closeButton: {
-    padding: 5,
-    marginLeft: 10
-  }
-});
